requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-cors==0.7.0
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
appdirs==1.4.4
asttokens==3.0.0
async-lru==2.0.4
atlassian-python-api==3.41.16
attrs==23.2.0
Babel==2.10.3
bcc==0.29.1
beartype==0.15.0
beautifulsoup4==4.12.3
bert-score==0.3.13
bleach==6.2.0
blinker==1.7.0
bm25s==0.2.6
bokeh==3.6.2
boto3==1.35.95
botocore==1.35.95
Brlapi==0.8.5
cachetools==5.5.0
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.1
chromedriver-autoinstaller==0.6.4
click==8.1.8
cloud-init==24.1.3
cohere==5.13.5
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
command-not-found==0.3
configobj==5.0.8
contourpy==1.3.1
cryptography==41.0.7
cssselect==1.2.0
cupshelpers==1.0
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.2.0
dbus-python==1.3.2
decorator==5.1.1
defer==1.0.6
Deprecated==1.2.15
dill==0.3.8
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
distro-info==1.7+build1
einops==0.8.0
executing==2.1.0
fastavro==1.10.0
fastembed==0.5.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.16.1
filetype==1.2.0
flatbuffers==24.12.23
fonttools==4.55.3
frozenlist==1.5.0
fs==2.4.16
fsspec==2024.9.0
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.0
google-api-python-client==2.157.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-pubsub==2.27.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grader==1.2
greenlet==3.1.1
groq==0.13.1
grpc-google-iam-v1==0.14.0
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
h3==4.1.2
html2text==2024.2.26
httpcore==1.0.7
httplib2==0.20.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.0
humanfriendly==10.0
idna==3.6
importlib_metadata==8.5.0
InstructorEmbedding==1.0.1
ipython==8.31.0
ipywidgets==8.1.5
jedi==0.19.2
jieba3k==0.35.1
Jinja2==3.1.2
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.0
jsonschema==4.10.3
jupyter_bokeh==4.0.5
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.28
langchain-experimental==0.3.4
langchain-groq==0.2.2
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.4
langsmith==0.2.7
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
linkify-it-py==2.0.3
llama-cloud==0.1.7
llama-hub==0.0.45
llama-index==0.12.9
llama-index-agent-coa==0.3.0
llama-index-agent-introspective==0.3.0
llama-index-agent-openai==0.4.1
llama-index-cli==0.4.0
llama-index-core==0.12.9
llama-index-embeddings-fastembed==0.3.0
llama-index-embeddings-huggingface==0.4.0
llama-index-embeddings-instructor==0.3.0
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.3
llama-index-llms-groq==0.3.1
llama-index-llms-openai==0.3.12
llama-index-llms-openai-like==0.3.3
llama-index-multi-modal-llms-openai==0.4.2
llama-index-postprocessor-cohere-rerank==0.3.0
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.2
llama-index-readers-llama-parse==0.4.0
llama-index-readers-web==0.3.3
llama-index-retrievers-bm25==0.5.0
llama-index-retrievers-pathway==0.3.0
llama-index-tools-graphql==0.3.0
llama-index-tools-tavily-research==0.3.0
llama-index-tools-wikipedia==0.3.0
llama-parse==0.5.19
loguru==0.7.3
louis==3.29.0
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
mmh3==4.1.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
narwhals==1.20.1
nest-asyncio==1.6.0
netaddr==0.8.0
netifaces==0.11.0
networkx==3.4.2
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.2.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
olefile==0.46
onnx==1.17.0
onnxruntime==1.20.1
openai==1.58.1
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
orjson==3.10.13
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
panel==1.5.5
param==2.2.0
parameterized==0.9.0
parso==0.8.4
pathway==0.16.4
pdfminer.six==20231228
pdfplumber==0.11.5
pexpect==4.9.0
phidata==2.7.6
pillow==10.4.0
playwright==1.49.1
prompt_toolkit==3.0.48
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
py_rust_stemmers==0.1.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycairo==1.25.1
pycups==2.0.1
pydantic==2.9.2
pydantic-settings==2.7.1
pydantic_core==2.23.4
pydeck==0.9.1
pyee==12.0.0
Pygments==2.17.2
PyGObject==3.48.2
PyJWT==2.7.0
PyMuPDF==1.25.1
pyparsing==3.1.1
pypdf==5.1.0
PyPDF2==3.0.1
pypdfium2==4.30.1
pyrsistent==0.20.0
pyserial==3.5
PySocks==1.7.1
PyStemmer==2.2.0.3
python-apt==2.7.7+ubuntu1
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-dotenv==1.0.1
python-sat==1.8.dev14
pytz==2024.1
pyviz_comms==3.0.3
pyxdg==0.28
PyYAML==6.0.1
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retrying==1.3.4
rich==13.7.1
rouge==1.0.1
rsa==4.9
s3transfer==0.10.4
safetensors==0.5.0
scikit-learn==1.6.0
scipy==1.15.0
selenium==4.27.1
sentence-transformers==2.7.0
setuptools==68.1.2
sgmllib3k==1.0.0
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spider-client==0.0.27
SQLAlchemy==2.0.36
sqlglot==10.6.1
stack-data==0.6.3
streamlit==1.41.1
striprtf==0.0.26
sympy==1.13.1
systemd-python==235
tavily-python==0.5.0
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tinysegmenter==0.3
tldextract==5.1.3
tokenizers==0.21.0
toml==0.10.2
tomli==2.2.1
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.1
trio==0.28.0
trio-websocket==0.11.1
triton==3.1.0
typer==0.15.1
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
uc-micro-py==1.0.3
ufw==0.36.2
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.0.7
wadllib==1.3.6
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.42.0
widgetsnbextension==4.0.13
wikipedia==1.4.0
wrapt==1.17.0
wsproto==1.2.0
xdg==5
xformers==0.0.29.post1
xkit==0.0.0
xxhash==3.5.0
xyzservices==2024.9.0
yarl==1.18.3
zipp==3.21.0
